requirements.txt
flask
numpy
opencv-python-headless
svgwrite
scikit-image
pillow
gunicorn
